Arreglado el menu de navegación en la web.  Andan los adcs.

+++ Webserver/Control/src/sensors.c
@@ -13,7 +13,7 @@ volatile uint8_t sensorNivelAgua = 0;
 volatile uint8_t sensorTemperatura = 15;
 volatile uint8_t sensorPh = 1;
 
-float sensorValue[] = {0, 15, 1};
+uint8_t sensorValue[] = {0, 15, 1};
 
 uint8_t sensorLimit[] = {30, 20, 25, 18, 8, 6};
